+++ src/helpers/functions/String.js
@@ -8,3 +8,10 @@ export function capitalize(str) {
     const regExp = /(?:[-_'])(?:[a-z0-9-_])/u;
     return str.replace(regExp, x => x.toUpperCase());
 }
+
+export function addTrailingSlash(str) {
+    if (typeof str !== "string") {
+        return str;
+    }
+    return str.slice(-1) !== "/" ? `${str}/` : str;
+}